Added testing file and moved fixtures
